# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Canary Islands Vacation Rental Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Property {
  id: number;
  name: string;
  location: string;
  description?: string;
  vvLicense: string;
  igicEnabled: boolean;
  nightly_rate: number;
  max_guests: number;
  icalImportUrls: string[];
  icalExportToken: string;
  lastSyncAt?: string | null;
  syncStatus?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePropertyBody {
  name: string;
  location: string;
  description?: string;
  vvLicense: string;
  igicEnabled: boolean;
  nightly_rate: number;
  max_guests: number;
  icalImportUrls?: string[];
}

export interface Booking {
  id: number;
  propertyId: number;
  guestName: string;
  guestEmail: string;
  guestPhone?: string | null;
  startDate: string;
  endDate: string;
  source: string;
  status: string;
  totalPrice?: number | null;
  igicAmount?: number | null;
  notes?: string | null;
  externalUid?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateBookingBody {
  propertyId: number;
  guestName: string;
  guestEmail: string;
  guestPhone?: string;
  startDate: string;
  endDate: string;
  source: string;
  status: string;
  totalPrice?: number;
  notes?: string;
}

export interface UpdateBookingBody {
  guestName?: string;
  guestEmail?: string;
  guestPhone?: string;
  startDate?: string;
  endDate?: string;
  source?: string;
  status?: string;
  totalPrice?: number;
  notes?: string;
}

export interface SyncResult {
  synced: number;
  conflicts: number;
  errors: string[];
  lastSyncAt: string;
}

export interface DashboardSummary {
  totalProperties: number;
  totalBookings: number;
  activeBookings: number;
  occupancyRate: number;
  revenueThisMonth: number;
  pendingCheckins: number;
}

export interface BookingWithProperty {
  id: number;
  propertyId: number;
  propertyName: string;
  propertyLocation: string;
  guestName: string;
  guestEmail: string;
  startDate: string;
  endDate: string;
  source: string;
  status: string;
  totalPrice?: number | null;
}

export interface OccupancyStat {
  month: string;
  occupancyRate: number;
  bookings: number;
}

export interface RevenueStat {
  source: string;
  revenue: number;
  bookings: number;
}

export type ExportPropertyIcalParams = {
  token: string;
};

export type ListBookingsParams = {
  propertyId?: number;
  status?: string;
};