# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Canary Islands Vacation Rental Management API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all properties
 */
export const ListPropertiesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  location: zod.string(),
  description: zod.string().optional(),
  vvLicense: zod.string(),
  igicEnabled: zod.boolean(),
  nightly_rate: zod.number(),
  max_guests: zod.number(),
  icalImportUrls: zod.array(zod.string()),
  icalExportToken: zod.string(),
  lastSyncAt: zod.string().nullish(),
  syncStatus: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListPropertiesResponse = zod.array(ListPropertiesResponseItem);

/**
 * @summary Create a property
 */
export const CreatePropertyBody = zod.object({
  name: zod.string(),
  location: zod.string(),
  description: zod.string().optional(),
  vvLicense: zod.string(),
  igicEnabled: zod.boolean(),
  nightly_rate: zod.number(),
  max_guests: zod.number(),
  icalImportUrls: zod.array(zod.string()).optional(),
});

/**
 * @summary Get a property by ID
 */
export const GetPropertyParams = zod.object({
  id: zod.coerce.number(),
});

export const GetPropertyResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  location: zod.string(),
  description: zod.string().optional(),
  vvLicense: zod.string(),
  igicEnabled: zod.boolean(),
  nightly_rate: zod.number(),
  max_guests: zod.number(),
  icalImportUrls: zod.array(zod.string()),
  icalExportToken: zod.string(),
  lastSyncAt: zod.string().nullish(),
  syncStatus: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update a property
 */
export const UpdatePropertyParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdatePropertyBody = zod.object({
  name: zod.string(),
  location: zod.string(),
  description: zod.string().optional(),
  vvLicense: zod.string(),
  igicEnabled: zod.boolean(),
  nightly_rate: zod.number(),
  max_guests: zod.number(),
  icalImportUrls: zod.array(zod.string()).optional(),
});

export const UpdatePropertyResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  location: zod.string(),
  description: zod.string().optional(),
  vvLicense: zod.string(),
  igicEnabled: zod.boolean(),
  nightly_rate: zod.number(),
  max_guests: zod.number(),
  icalImportUrls: zod.array(zod.string()),
  icalExportToken: zod.string(),
  lastSyncAt: zod.string().nullish(),
  syncStatus: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a property
 */
export const DeletePropertyParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Trigger iCal sync for a property
 */
export const SyncPropertyIcalParams = zod.object({
  id: zod.coerce.number(),
});

export const SyncPropertyIcalResponse = zod.object({
  synced: zod.number(),
  conflicts: zod.number(),
  errors: zod.array(zod.string()),
  lastSyncAt: zod.string(),
});

/**
 * @summary Export iCal feed for a property
 */
export const ExportPropertyIcalParams = zod.object({
  id: zod.coerce.number(),
});

export const ExportPropertyIcalQueryParams = zod.object({
  token: zod.coerce.string(),
});

/**
 * @summary List bookings
 */
export const ListBookingsQueryParams = zod.object({
  propertyId: zod.coerce.number().optional(),
  status: zod.coerce.string().optional(),
});

export const ListBookingsResponseItem = zod.object({
  id: zod.number(),
  propertyId: zod.number(),
  guestName: zod.string(),
  guestEmail: zod.string(),
  guestPhone: zod.string().nullish(),
  startDate: zod.string(),
  endDate: zod.string(),
  source: zod.string(),
  status: zod.string(),
  totalPrice: zod.number().nullish(),
  igicAmount: zod.number().nullish(),
  notes: zod.string().nullish(),
  externalUid: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListBookingsResponse = zod.array(ListBookingsResponseItem);

/**
 * @summary Create a booking
 */
export const CreateBookingBody = zod.object({
  propertyId: zod.number(),
  guestName: zod.string(),
  guestEmail: zod.string(),
  guestPhone: zod.string().optional(),
  startDate: zod.string(),
  endDate: zod.string(),
  source: zod.string(),
  status: zod.string(),
  totalPrice: zod.number().optional(),
  notes: zod.string().optional(),
});

/**
 * @summary Get a booking by ID
 */
export const GetBookingParams = zod.object({
  id: zod.coerce.number(),
});

export const GetBookingResponse = zod.object({
  id: zod.number(),
  propertyId: zod.number(),
  guestName: zod.string(),
  guestEmail: zod.string(),
  guestPhone: zod.string().nullish(),
  startDate: zod.string(),
  endDate: zod.string(),
  source: zod.string(),
  status: zod.string(),
  totalPrice: zod.number().nullish(),
  igicAmount: zod.number().nullish(),
  notes: zod.string().nullish(),
  externalUid: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update a booking
 */
export const UpdateBookingParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateBookingBody = zod.object({
  guestName: zod.string().optional(),
  guestEmail: zod.string().optional(),
  guestPhone: zod.string().optional(),
  startDate: zod.string().optional(),
  endDate: zod.string().optional(),
  source: zod.string().optional(),
  status: zod.string().optional(),
  totalPrice: zod.number().optional(),
  notes: zod.string().optional(),
});

export const UpdateBookingResponse = zod.object({
  id: zod.number(),
  propertyId: zod.number(),
  guestName: zod.string(),
  guestEmail: zod.string(),
  guestPhone: zod.string().nullish(),
  startDate: zod.string(),
  endDate: zod.string(),
  source: zod.string(),
  status: zod.string(),
  totalPrice: zod.number().nullish(),
  igicAmount: zod.number().nullish(),
  notes: zod.string().nullish(),
  externalUid: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a booking
 */
export const DeleteBookingParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get dashboard summary stats
 */
export const GetDashboardSummaryResponse = zod.object({
  totalProperties: zod.number(),
  totalBookings: zod.number(),
  activeBookings: zod.number(),
  occupancyRate: zod.number(),
  revenueThisMonth: zod.number(),
  pendingCheckins: zod.number(),
});

/**
 * @summary Get upcoming bookings (next 30 days)
 */
export const GetUpcomingBookingsResponseItem = zod.object({
  id: zod.number(),
  propertyId: zod.number(),
  propertyName: zod.string(),
  propertyLocation: zod.string(),
  guestName: zod.string(),
  guestEmail: zod.string(),
  startDate: zod.string(),
  endDate: zod.string(),
  source: zod.string(),
  status: zod.string(),
  totalPrice: zod.number().nullish(),
});
export const GetUpcomingBookingsResponse = zod.array(
  GetUpcomingBookingsResponseItem,
);

/**
 * @summary Get monthly occupancy stats
 */
export const GetOccupancyStatsResponseItem = zod.object({
  month: zod.string(),
  occupancyRate: zod.number(),
  bookings: zod.number(),
});
export const GetOccupancyStatsResponse = zod.array(
  GetOccupancyStatsResponseItem,
);

/**
 * @summary Get revenue breakdown by source
 */
export const GetRevenueStatsResponseItem = zod.object({
  source: zod.string(),
  revenue: zod.number(),
  bookings: zod.number(),
});
export const GetRevenueStatsResponse = zod.array(GetRevenueStatsResponseItem);